Cultural Sensitivity
Okay, this is going to make me look at bit (more) stupid - but whilst I'm sure the Bacon ipsum was put in there purely for fun and without any malice, it's important to consider the feelings and sensitivities of your users, developers & customers. 

I love a bacon butty as much as the next person, but I conscious that many cultures and religions aren't fond of pork. So the existing text goes a little (well, very!) overboard in that respect.